+++ CHANGELOG.md
@@ -1,6 +1,31 @@
 # CraftHelper updates
 
-## 1.6.6 (XX.02.2023)
+# 1.7 (05.03.2023)
+
+### New features
+
+- Added option for delayed search (Need to press Enter to perform search) (by default OFF)
+- Added amounts of items in each category to category selector
+- Added paste function to search bar (Right click)
+
+### Improvements
+
+- Optimized rendering and improved overall performance
+- Removed vanilla recipes that require Anvil nearby
+- Added Traditional Chinese translation (thanks to [TW_Billy](https://steamcommunity.com/profiles/76561198001484294))
+- Types and categories are now filtered along with search query
+- Use RecipeManager.getSourceItemsAll for each source instead of getAvailableItemsAll when collecting available items (incorrect counts otherwise, needs testing)
+- Changed event used to load all recipes (OnGameStart -> OnLoad)
+
+### Bugfixes
+
+- Fixed incorrect crafting when [Proximity Inventory](https://steamcommunity.com/sharedfiles/filedetails/?id=2847184718) active (extra loot tab not accounted for)
+- Fixed recipe type amounts (valid/invalid) not updating properly
+- Fixed logic in CHC_main.common.isRecipeValid
+- Fixed logical errors with types/categories
+- Various UI fixes
+
+## 1.6.6 (23.02.2023)
 
 ### New features